fix(2fa): guard the verify page against direct access
The /2fa verify page was a pure client component with no server-side
guard, so visiting it directly just sat there instead of redirecting.

Turn it into a server component that redirects before rendering:
- a full session means 2FA is already satisfied (or not required) -> /
- no session and no pending 2FA challenge -> /login
- only a genuine pending challenge keeps the page

The pending state is detected via a new hasPendingTwoFactor() guard in
lib/auth-guards.ts, which checks for better-auth's short-lived signed
`two_factor` cookie (set while a challenge is in flight, when the full
session cookie is absent). The cookie name is derived from better-auth's
own cookie helper so the __Secure- prefix stays correct in production.

The verify form is extracted into two-factor-verify.tsx, mirroring the
existing 2fa/setup page structure.
